still working on buy crypto screen amount input, validations, converts and code cleanup of crypto purchase flow

+++ data/src/main/java/com/multimoney/data/networking/GraphqlApi.kt
@@ -1564,7 +1564,7 @@ class GraphqlApi @Inject constructor(
         market: String,
         commissionAmount: Double,
         taxAmount: Double,
-        accountToken: Double,
+        accountToken: Long,
         exchangeRate: Double,
         idBrand: Int,
         user: String,
